Add 16-bit CPU instructions, GPU start

+++ src/emulator/VideoOutput.ts
@@ -1,5 +1,8 @@
 interface VideoOutput {
-    receive(data: Uint8ClampedArray): void;
+    /**
+     * Receives the GameBoy's output, as an array of RGBA values.
+     */
+    receive(data: Uint32Array): void;
 }
 
 export default VideoOutput;
